requirements.txt
firebase-admin==6.2.0
google-cloud-firestore==2.11.1
google-auth==2.23.4
google-api-core==2.11.1 